% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{complete_igrtcm}
\alias{complete_igrtcm}
\title{Log-likelihood and gradient of the complete GRTC Model on one observation}
\usage{
complete_igrtcm(
  THETA,
  EXAMS_GRADES,
  EXAMS_DAYS,
  EXAMS_SET,
  EXAMS_OBSFLAG,
  MAX_DAY,
  N_GRADES,
  N_EXAMS,
  ABILITY,
  SPEED
)
}
\arguments{
\item{THETA}{Parameter vector}

\item{EXAMS_GRADES}{Vector of exam grades}

\item{EXAMS_DAYS}{Vector of exam days}

\item{EXAMS_SET}{Vector of binary values. 1 to include an exam in the study-plan, 0 otherwise}

\item{EXAMS_OBSFLAG}{Vector of binary values. 1 if the exam is observed, 0 otherwise}

\item{MAX_DAY}{Maximum day of observation}

\item{N_GRADES}{Number of possible grades modelled}

\item{N_EXAMS}{Number of possible exams modelled}

\item{ABILITY}{Ability latent parameter}

\item{SPEED}{Speed latent parameter}
}
\description{
Used for internal testing
}
