% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP.R
\name{compute_map}
\alias{compute_map}
\title{MAP GRTCM}
\usage{
compute_map(FIT, MATSTART = NULL, TIDY = TRUE)
}
\arguments{
\item{FIT}{Output from \link{fit_EM} or \link{fit_BFGS}}

\item{MATSTART}{Starting values for the latent scores}

\item{TIDY}{Return tidy parameter table}
}
\description{
Compute MAP scores for Graded Response Time Censored Model
}
