% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRGRTCM.R
\name{compute_se}
\alias{compute_se}
\title{Compute standard errors from GRTCM fit}
\usage{
compute_se(FIT, TIDY = TRUE)
}
\arguments{
\item{FIT}{Output from \link{fit_CRGRTCM}}

\item{TIDY}{Return tidy parameter table}
}
\description{
Compute standard errors from GRTCM fit
}
