% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{compute_stderr}
\alias{compute_stderr}
\title{Compute standard errors from GRTCM fit}
\usage{
compute_stderr(FIT, NUMDER = FALSE, TIDY = TRUE, GRID = NULL, WEIGHTS = NULL)
}
\arguments{
\item{FIT}{Output from \link{fit_EM} or \link{fit_BFGS}}

\item{TIDY}{Return tidy parameter table}
}
\description{
Compute standard errors from GRTCM fit
}
