% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRTCM.R
\name{eap_GRTCM}
\alias{eap_GRTCM}
\title{EAP GRTCM}
\usage{
eap_GRTCM(FIT, TIDY = TRUE)
}
\arguments{
\item{FIT}{Output from \link{fit_GRTCM}}

\item{TIDY}{Return tidy parameter table}
}
\description{
Compute EAP scores for Graded Response Time Censored Model
}
