% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFGS.R
\name{fit_BFGS}
\alias{fit_BFGS}
\title{Fit CRGRTCM}
\usage{
fit_BFGS(
  DATA,
  GRADE_COL = "xcat",
  EXAM_COL = "itemID",
  STUDID_COL = "subjectID",
  TIME_COL = "time",
  MAXTIME_COL = "maxTime",
  GRID,
  WEIGHTS,
  THETA_START = NULL,
  MOD,
  ...
)
}
\arguments{
\item{DATA}{Dataset in long form}

\item{GRADE_COL}{Label of the column where to find the grades}

\item{EXAM_COL}{Label of the column where to find exams ids}

\item{STUDID_COL}{Label of the column where to find students ids.}

\item{TIME_COL}{Label of the column where to find exams times.}

\item{MAXTIME_COL}{Label of the column where to find the maximum time observed for each student.}

\item{GRID}{Grid of quadrature points to be used.}

\item{WEIGHTS}{Weights for quadrature points}

\item{THETA_START}{Optional starting parameter vector}

\item{...}{additional arguments to be passed to \code{\link[ucminf:ucminf]{ucminf::ucminf()}}}
}
\description{
Fit Competing Risk Graded Response Time Censored Model
}
