% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRTCM.R
\name{fit_GRTCM}
\alias{fit_GRTCM}
\title{Fit GRTCM}
\usage{
fit_GRTCM(
  DATA,
  GRADE_COL = "xcat",
  EXAM_COL = "itemID",
  STUDID_COL = "subjectID",
  TIME_COL = "time",
  MAXTIME_COL = "maxTime",
  GRID,
  WEIGHTS,
  THETA_START = NULL
)
}
\arguments{
\item{DATA}{Dataset in long form}

\item{GRADE_COL}{Label of the column where to find the grades}

\item{EXAM_COL}{Label of the column where to find exams ids}

\item{STUDID_COL}{Label of the column where to find students ids.}

\item{TIME_COL}{Label of the column where to find exams times.}

\item{MAXTIME_COL}{Label of the column where to find the maximum time observed for each student.}

\item{GRID}{Grid of quadrature points to be used.}

\item{WEIGHTS}{Weights for quadrature points}

\item{THETA_START}{Optional starting parameter vector}
}
\description{
Fit Graded Response Time Censored Model
}
