% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_fitted_outcome}
\alias{get_fitted_outcome}
\title{get fitted outcome probabilities given exam performances}
\usage{
get_fitted_outcome(FIT)
}
\arguments{
\item{FIT}{Output from \link{fit_EM} or \link{fit_BFGS}}
}
\description{
get fitted outcome probabilities given exam performances
}
