% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOGLIK.R
\name{get_loglik}
\alias{get_loglik}
\title{Compute Marginal Loglik}
\usage{
get_loglik(FIT, GRID = NULL, WEIGHTS = NULL, INIT = FALSE)
}
\arguments{
\item{FIT}{Output from \link{fit_EM} or \link{fit_BFGS}}

\item{GRID}{Grid of quadrature points to be used.}

\item{WEIGHTS}{Weights for quadrature points.}

\item{INIT}{If TRUE, it evaluates the marginal log likelihood at the initial parameter values}
}
\description{
Compute Marginal Loglik
}
