% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GRTCM_complete_obs}
\alias{GRTCM_complete_obs}
\title{IRT complete observation}
\usage{
GRTCM_complete_obs(
  THETA,
  EXAMS_GRADES,
  EXAMS_DAYS,
  EXAMS_OBSFLAG,
  EXAMS_SET,
  MAX_DAY,
  N_GRADES,
  N_EXAMS,
  ABILITY,
  SPEED
)
}
\arguments{
\item{THETA}{Suitable parameter vector as provided by \link{parList2Vec}}

\item{EXAMS_GRADES}{Vector of grades.}

\item{EXAMS_DAYS}{Vector of times.}

\item{EXAMS_OBSFLAG}{Vector of booleans.\code{TRUE} elements represent observed exams. \code{FALSE} elements the unobserved ones.}

\item{EXAMS_SET}{Vector filled with booleans.\code{TRUE} elements represent exams in the study plan. \code{FALSE} elements non-relevant ones.}

\item{MAX_DAY}{Last day of observation}

\item{N_GRADES}{Number of grades modelled.}

\item{N_EXAMS}{Number of exams modelled}

\item{ABILITY}{Ability value.}

\item{SPEED}{Speed value.}
}
\value{
It returns the value of the integrand function,
given the parameters and the data of a single observation.
}
\description{
Evaluate the IRT complete model log likelihood of a single observation
}
