% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GRTCM_GH}
\alias{GRTCM_GH}
\title{GRTCM computations with Gaussian quadrature}
\usage{
GRTCM_GH(
  THETA,
  EXAMS_GRADES,
  EXAMS_DAYS,
  EXAMS_SET,
  EXAMS_OBSFLAG,
  MAX_DAY,
  GRID,
  WEIGHTS,
  N_GRADES,
  N_EXAMS,
  GRFLAG = TRUE,
  ROTGRID = TRUE
)
}
\arguments{
\item{THETA}{Parameter vector}

\item{EXAMS_GRADES}{Matrix of exam grades}

\item{EXAMS_DAYS}{Matrix of exam days}

\item{EXAMS_SET}{Matrix of binary values. 1 to include an exam in the study-plan, 0 otherwise}

\item{EXAMS_OBSFLAG}{Matrix of binary values. 1 if the exam is observed, 0 otherwise}

\item{MAX_DAY}{Vector with maximum day per observation}

\item{GRID}{Grid of quadrature points}

\item{WEIGHTS}{Weights of quadrature points}

\item{N_GRADES}{Number of possible grades modelled}

\item{N_EXAMS}{Number of possible exams modelled}

\item{GRFLAG}{Set to true to return the gradient along the log-likelihood}

\item{ROTGRID}{TRUE to rotate the quadrature grid using the latent covariance matrix}
}
\description{
Compute log-likelihood an gradient of the Graded response time-censored model
}
