% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{irtMat2Vec}
\alias{irtMat2Vec}
\title{IRT parameters to matrix}
\usage{
irtMat2Vec(MAT)
}
\arguments{
\item{MAT}{Parameter matrix (as returned by \link{irtVec2Mat}).}
}
\value{
It reverts the transformation applied by \link{irtVec2Mat}
}
\description{
From constrained parameter matrix to unconstrained parameter vector
}
