% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{irtVec2Mat}
\alias{irtVec2Mat}
\title{IRT parameters to matrix}
\usage{
irtVec2Mat(THETA_IRT, N_GRADES, N_EXAMS, LABS_EXAMS = NULL, LABS_GRADES = NULL)
}
\arguments{
\item{THETA_IRT}{Parameter vector.}

\item{N_GRADES}{number of grades modeled.}

\item{N_EXAMS}{number of exams.}

\item{LABS_EXAMS}{optional label for exams}

\item{LABS_GRADES}{optional label for grades#'}
}
\description{
Reparametrise IRT parameters and rearrange them in a matrix
}
