% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRGRTCM.R
\name{map_CRGRTCM}
\alias{map_CRGRTCM}
\title{MAP GRTCM}
\usage{
map_CRGRTCM(FIT, MATSTART = NULL, TIDY = TRUE)
}
\arguments{
\item{FIT}{Output from \link{fit_CRGRTCM}}

\item{MATSTART}{Starting values for the latent scores}

\item{TIDY}{Return tidy parameter table}
}
\description{
Compute MAP scores for Graded Response Time Censored Model
}
