% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRTCM.R
\name{map_GRTCM}
\alias{map_GRTCM}
\title{MAP GRTCM}
\usage{
map_GRTCM(FIT, TIDY = TRUE)
}
\arguments{
\item{FIT}{Output from \link{fit_GRTCM}}

\item{TIDY}{Return tidy parameter table}
}
\description{
Compute MAP scores for Graded Response Time Censored Model
}
