% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{parList2Vec}
\alias{parList2Vec}
\title{Parameters list to unconstrained vector}
\usage{
parList2Vec(LIST)
}
\arguments{
\item{LIST}{List of parameters as returned by \link{parVec2List}}
}
\description{
Parameters list to unconstrained vector
}
