% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{parVec2List}
\alias{parVec2List}
\title{Unconstrained parameter vector to list}
\usage{
parVec2List(THETA, N_GRADES, N_EXAMS, LABS_EXAMS = NULL, LABS_GRADES = NULL)
}
\arguments{
\item{THETA}{Parameter vector.}

\item{N_GRADES}{number of grades modeled.}

\item{N_EXAMS}{number of exams.}

\item{LABS_EXAMS}{optional label for exams}

\item{LABS_GRADES}{optional label for grades#'}
}
\description{
Reparametrise unconstrained parameters and rearrange them into a list
}
