% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reparThr}
\alias{reparThr}
\title{Intercepts reparameterisation}
\usage{
reparThr(X, CON2UN = FALSE)
}
\arguments{
\item{X}{Intercepts or unconstrained parameter values for grades low-to-high.}

\item{CON2UN}{TRUE if going from the constrained space to the unconstrained one.}
}
\value{
It allows to go back and forth from constrained intercepts
to unconstrained parameters.
}
\description{
Intercepts reparameterisation
}
