% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRTCM.R
\name{se_GRTCM}
\alias{se_GRTCM}
\title{Compute standard errors from GRTCM fit}
\usage{
se_GRTCM(FIT, TIDY = TRUE)
}
\arguments{
\item{FIT}{Output from \link{fit_GRTCM}}

\item{TIDY}{Return tidy parameter table}
}
\description{
Compute standard errors from GRTCM fit
}
