% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{parVec2Repar}
\alias{parVec2Repar}
\title{Unconstrained parameter vector to reparametrisation}
\usage{
parVec2Repar(
  THETA,
  N_GRADES,
  N_EXAMS,
  N_COV,
  YB,
  LABS_EXAMS = NULL,
  LABS_GRADES = NULL,
  LABS_COV = NULL,
  TIDY = FALSE
)
}
\arguments{
\item{THETA}{Parameter vector.}

\item{N_GRADES}{number of grades modeled.}

\item{N_EXAMS}{number of exams.}

\item{N_COV}{number of covariates in the competing risk model}

\item{YB}{Maximum number of years before graduation}

\item{LABS_EXAMS}{optional label for exams}

\item{LABS_GRADES}{optional label for grades}

\item{LABS_COV}{optional labels for covariates in the competing risk model}

\item{TIDY}{TRUE to get tidy parameters table}
}
\description{
Reparametrise unconstrained parameters
}
